+++ R/manystates-package.R
@@ -1,6 +1,10 @@
 #' @keywords internal
 "_PACKAGE"
 
+#' @importFrom manydata consolidate
+#' @export
+manydata::consolidate
+
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start