+++ R/manystates-states.R
@@ -1,31 +1,52 @@
-#' states database documentation
-#'
-#' @format The states database is a list that contains the
-#' following 7 datasets: GW, ISD, COW, ICOW, ICOW_COL, RATRULES, HUGGO_STATES.
-#' For more information and references to each of the datasets used,
-#' please use the `data_source()` and `data_contrast()` functions.
-#'\describe{
-#' \item{GW: }{A dataset with 216 observations and the following
-#' 5 variables: cowID, Beg, End, cowNR, StateName.}
-#' \item{ISD: }{A dataset with 362 observations and the following
-#' 7 variables: cowID, Beg, End, cowNR, StateName, Micro, NewState.}
-#' \item{COW: }{A dataset with 243 observations and the following
-#' 5 variables: cowID, Beg, End, cowNR, StateName.}
-#' \item{ICOW: }{A dataset with 217 observations and the following
-#' 15 variables: cowID, StateName, ColRuler, IndFrom, IndDate, IndViol, IndType,
-#' SecFrom, SecDate, SecViol, Into, IntoDate, COWsys, GWsys, Notes.}
-#' \item{ICOW_COL: }{A dataset with 221 observations and the following
-#' 9 variables: cowID, cowNR, StateName, Beg, cowID_Origin, IndepType, Beg_COW,
-#' Beg_GW, Beg_Polity2.}
-#' \item{RATRULES: }{A dataset with 177 observations and the following
-#' 3 variables: stateID, StateName, RatProcedure.}
-#' \item{HUGGO_STATES: }{A dataset with 574 observations and the following
-#' 12 variables: stateID, StateName, Capital, Beg, End, Latitude, Longitude,
-#' Area, Region, RatProcedure, Constitutional Description, Source.}
+#' States datacube
+#' @description The `states` datacube is a list containing 
+#'   `r length(manystates::states)` datasets: 
+#'   `r cli::pluralize("{names(manystates::states)}")`.
+#'   It is a work-in-progress, so please let us know if you have any comments or suggestions.
+#' @format
+#' \describe{
+#' \item{HUGGO: }{A dataset with `r prettyNum(nrow(manystates::states$HUGGO), big.mark=",")` 
+#' observations and `r ncol(manystates::states$HUGGO)` variables: 
+#' `r cli::pluralize("{names(manystates::states$HUGGO)}")`.}
+#' \item{GW: }{A dataset with `r prettyNum(nrow(manystates::states$GW), big.mark=",")` 
+#' observations and `r ncol(manystates::states$GW)` variables: 
+#' `r cli::pluralize("{names(manystates::states$GW)}")`.}
+#' \item{ISD: }{A dataset with `r prettyNum(nrow(manystates::states$ISD), big.mark=",")` 
+#' observations and `r ncol(manystates::states$ISD)` variables: 
+#' `r cli::pluralize("{names(manystates::states$ISD)}")`.}
 #' }
-#'
+#' For more information and references to each of the datasets used,
+#' please use the `manydata::call_sources()` and `manydata::compare_dimensions()` functions.
+#' @source
+#' \describe{
+#' \item{HUGGO: }{
+#' Hollway, James, Henrique Sposito, and Jael Tan. 2021. _manystates: States for manydata_.
+#' }
+#' \item{GW: }{
+#' Gleditsch, K.S., and M. D. Ward. “Interstate system membership: A revised list of the independent states since 1816”.
+#' _International Interactions_ 25.4 (1999), pp. 393-413.
+#' \url{http://ksgleditsch.com/data-4.html}
+#' }
+#' \item{ISD: }{
+#' Griffiths, R.D., and C. R. Butcher. “Introducing the international system (s) dataset (ISD), 1816-2011”.
+#' _International Interactions_ 39.5 (2013), pp. 748-768.
+#' \url{http://www.ryan-griffiths.com/data}
+#' }
+#' }
+#' @section Mapping:
+#' 
+#' |  *manystates*  | *GW*  | *ISD* | 
+#' |:---------------|:------|:------|
+#' | stateID  | | |
+#' | Begin | Start | Start |
+#' | End | Finish | Finish |
+#' | StateName | Name of State | State.Name |
+#' | cowID | Cow ID | COW.ID |
+#' | cowNR | Cow NR. | COW.Nr |
+#' 
+#' @md
 #' @details
 #' ``` {r, echo = FALSE, warning = FALSE}
-#' lapply(states, messydates::mreport)
+#' lapply(states, manydata::mreport)
 #' ```
 "states"