+++ data-raw/contiguity/HUGGO_CONT/prepare-HUGGO_CONT.R
@@ -15,22 +15,50 @@ HUGGO_CONT <- readr::read_csv("data-raw/contiguity/HUGGO_CONT/FAO and Region Mem
 HUGGO_CONT <- as_tibble(HUGGO_CONT) %>%
   dplyr::filter(ID != "ISO3") %>%
   # filtering removes the rows that contain repetitions of variable names only
-  dplyr::rename(stateID = ID, EntityType = CATEGORY,
-                FAOmembership = FAO_MEMBERS, Group = IS_IN_GROUP, url = URI) %>%
-  manydata::transmutate(StateName = manypkgs::standardise_titles(LISTNAME_EN),
-                        Contiguity = manypkgs::standardise_titles(HAS_BORDER_WITH),
-                        Beg = messydates::as_messydate(VALID_SINCE),
+  dplyr::rename(stateID = ID, url = URL) %>%
+  manydata::transmutate(StateName1 = manypkgs::standardise_titles(LISTNAME_EN),
+                        Contiguity = HAS_BORDER_WITH,
+                        Begin = messydates::as_messydate(VALID_SINCE),
                         End = messydates::as_messydate(VALID_UNTIL)) %>%
-  dplyr::select(stateID, StateName, Beg, End, Contiguity, EntityType,
-                FAOmembership, Group, url) %>%
-  dplyr::arrange(Beg, stateID)
+  dplyr::mutate(StateName1 = manypkgs::code_states(StateName1, activity = F,
+                                                  replace = "names"),
+                StateName1 = ifelse(stateID == "CIV", "Cote d'Ivoire",
+                                   StateName1),
+                StateName1 = ifelse(stringr::str_detect(StateName1, "Korea - "),
+                                    "Democratic People's Republic of Korea",
+                                    StateName1)) %>%
+  tidyr::separate_wider_delim(Contiguity, delim = ",",
+                              names_sep = "_", too_few = "align_start") %>%
+  tidyr::pivot_longer(c("Contiguity_1":"Contiguity_14"),
+                      values_to = "StateName", values_drop_na = TRUE) %>%
+  dplyr::mutate(StateName = stringr::str_remove_all(StateName, "_the"),
+                StateName = stringr::str_replace_all(StateName, "_", " "),
+                StateName2 = manypkgs::code_states(StateName, activity = F,
+                                                   replace = "names"),
+                StateName2 = ifelse(is.na(StateName2), StateName, StateName2),
+                StateName2 = ifelse(stringr::str_detect(StateName2, "Korea - "),
+                                    "Democratic People's Republic of Korea",
+                                    StateName2),
+                stateID2 = manypkgs::code_states(StateName2, activity = F,
+                                                 replace = "ID"),
+                stateID2 = ifelse(stateID2 == "KOR - PRK", "PRK", stateID2),
+                stateID1 = manypkgs::code_states(StateName1, activity = F,
+                                                 replace = "ID"),
+                stateID1 = ifelse(is.na(stateID1), stateID, stateID1),
+                stateID1 = ifelse(stateID1 == "KOR - PRK", "PRK", stateID1),
+                # coding standardised with COW_CONT: 1 = shared border
+                ContiguityType = 1) %>%
+  dplyr::select(stateID1, stateID2, Begin, End, StateName1, StateName2,
+                ContiguityType, url) %>%
+  dplyr::arrange(Begin, stateID1)
 
 # make sure all vars are correctly coded as NA if necessary
 HUGGO_CONT <- HUGGO_CONT %>% 
   dplyr::mutate(across(everything(),
                        ~stringr::str_replace_all(., "^NA$", NA_character_))) %>%
-  dplyr::mutate(Beg = messydates::as_messydate(Beg),
-         End = messydates::as_messydate(End)) %>% 
+  dplyr::mutate(Begin = messydates::as_messydate(Begin),
+                End = messydates::as_messydate(End),
+                ContiguityType = as.numeric(ContiguityType)) %>% 
   dplyr::distinct(.keep_all = TRUE)
 
 # manypkgs includes several functions that should help cleaning
@@ -55,5 +83,5 @@ HUGGO_CONT <- HUGGO_CONT %>%
 # that you're including in the package.
 # To add a template of .bib file to the package,
 # please run `manypkgs::add_bib("contiguity", "HUGGO_CONT")`.
-manypkgs::export_data(HUGGO_CONT, database = "contiguity",
+manypkgs::export_data(HUGGO_CONT, datacube = "contiguity",
                       URL = "Hand-coded data by the GGO team")